moved contributor status to personal information
